moved zign stuff under zignatures.c
