+++ librz/core/core.c
@@ -302,6 +302,10 @@ static bool __syncDebugMaps(RzCore *core) {
 	return false;
 }
 
+static const RzList *__flagsGet(RzCore *core, ut64 offset) {
+	return rz_flag_get_list(core->flags, offset);
+}
+
 RZ_API int rz_core_bind(RzCore *core, RzCoreBind *bnd) {
 	bnd->core = core;
 	bnd->bphit = (RzCoreDebugBpHit)rz_core_debug_breakpoint_hit;
@@ -320,6 +324,7 @@ RZ_API int rz_core_bind(RzCore *core, RzCoreBind *bnd) {
 	bnd->numGet = (RzCoreNumGet)numget;
 	bnd->isMapped = (RzCoreIsMapped)__isMapped;
 	bnd->syncDebugMaps = (RzCoreDebugMapsSync)__syncDebugMaps;
+	bnd->flagsGet = (RzCoreFlagsGet)__flagsGet;
 	return true;
 }
